requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-genai==0.4.0
redis[hiredis]==5.2.1
asyncpg==0.30.0
python-dotenv==1.0.1
pydantic==2.10.4
# Database
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
alembic==1.14.1
# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18
# Reporting
openpyxl==3.1.5
reportlab==4.2.5
# Notifications
aiosmtplib==3.0.2
# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
# Deployment
gunicorn==23.0.0
